requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core
azure-functions
azure-identity
azure-keyvault-secrets
azure-storage-blob
#adlfs
#fsspec
geopandas
psycopg2-binary==2.9.9
rasterio==1.3.10
requests
rio-cogeo
fiona
